refactored performance data manager to not rely on Game Manager, added some logic to make the level_fr still work
