+++ Assets/ML-Agents/Timers/End_Screen_timers.json
@@ -1,1 +1,1 @@
-{"count":1,"self":15.7244096,"total":15.7663599,"children":{"AgentSendState":{"count":529,"self":0.022973,"total":0.0250005,"children":{"CollectObservations":{"count":82,"self":0.0020275,"total":0.0020275,"children":null},"CollectDiscreteActionMasks":{"count":82,"self":0,"total":0,"children":null}}},"DecideAction":{"count":529,"self":0.0079849,"total":0.0079849,"children":null},"AgentAct":{"count":529,"self":0.0079645,"total":0.0079645,"children":null}},"gauges":{"Paddle.CumulativeReward":{"count":1,"max":-9.5,"min":-9.5,"runningAverage":-9.5,"value":-9.5,"weightedAverage":-9.5}},"metadata":{"timer_format_version":"0.1.0","start_time_seconds":"1613438045","unity_version":"2019.4.18f1","command_line_arguments":"C:\\Program Files\\Unity\\Hub\\Editor\\2019.4.18f1\\Editor\\Unity.exe -projectpath C:\/Users\/frann\/Documents\/UnityProjects\/CS467\/analysis_logging -useHub -hubIPC -cloudEnvironment production -hubSessionId 35b826e0-6fef-11eb-8804-59ce1fd46b66 -accessToken BDdcbCpNTfX5zDOGnVxwAP4MYYOa0xD8GqTTGdXLwDU002f","communication_protocol_version":"1.0.0","com.unity.ml-agents_version":"1.0.6","scene_name":"End_Screen","end_time_seconds":"1613438060"}}
+{"count":1,"self":8.8151488,"total":8.8261497,"children":{"AgentSendState":{"count":197,"self":0.0049997999999999996,"total":0.0059995,"children":{"CollectObservations":{"count":35,"self":0.00099969999999999985,"total":0.00099969999999999985,"children":null},"CollectDiscreteActionMasks":{"count":35,"self":0,"total":0,"children":null}}},"DecideAction":{"count":197,"self":0.0030011,"total":0.0030011,"children":null},"AgentAct":{"count":197,"self":0.0010006,"total":0.0010006,"children":null}},"gauges":{"Paddle.CumulativeReward":{"count":1,"max":-9.8,"min":-9.8,"runningAverage":-9.8,"value":-9.8,"weightedAverage":-9.8}},"metadata":{"timer_format_version":"0.1.0","start_time_seconds":"1613527849","unity_version":"2019.4.18f1","command_line_arguments":"C:\\Program Files\\Unity\\Hub\\Editor\\2019.4.18f1\\Editor\\Unity.exe -projectpath C:\/Users\/frann\/Documents\/UnityProjects\/CS467\/analysis_logging -useHub -hubIPC -cloudEnvironment production -hubSessionId c9277360-70a6-11eb-9363-db6cc44d73e6 -accessToken D6fAzro6ock-1efwinfDKR7nsoVY-KpOvxwKGxE8CQY002f","communication_protocol_version":"1.0.0","com.unity.ml-agents_version":"1.0.6","scene_name":"End_Screen","end_time_seconds":"1613527857"}}
