+++ Assets/ML-Agents/Timers/Level_FR_timers.json
@@ -1,1 +1,1 @@
-{"count":1,"self":74.276896,"total":74.3049757,"children":{"AgentSendState":{"count":626,"self":0.0134813,"total":0.0134813,"children":{"CollectObservations":{"count":61,"self":0,"total":0,"children":null},"CollectDiscreteActionMasks":{"count":61,"self":0,"total":0,"children":null}}},"DecideAction":{"count":626,"self":0.0069809999999999994,"total":0.0069809999999999994,"children":null},"AgentAct":{"count":626,"self":0.0076171999999999993,"total":0.0076171999999999993,"children":null}},"gauges":{"Paddle.CumulativeReward":{"count":1,"max":-9.7,"min":-9.7,"runningAverage":-9.7,"value":-9.7,"weightedAverage":-9.7}},"metadata":{"timer_format_version":"0.1.0","start_time_seconds":"1613437969","unity_version":"2019.4.18f1","command_line_arguments":"C:\\Program Files\\Unity\\Hub\\Editor\\2019.4.18f1\\Editor\\Unity.exe -projectpath C:\/Users\/frann\/Documents\/UnityProjects\/CS467\/analysis_logging -useHub -hubIPC -cloudEnvironment production -hubSessionId 35b826e0-6fef-11eb-8804-59ce1fd46b66 -accessToken BDdcbCpNTfX5zDOGnVxwAP4MYYOa0xD8GqTTGdXLwDU002f","communication_protocol_version":"1.0.0","com.unity.ml-agents_version":"1.0.6","scene_name":"Level_FR","end_time_seconds":"1613438043"}}
+{"count":1,"self":5.879842,"total":5.8899247,"children":null,"gauges":{},"metadata":{"timer_format_version":"0.1.0","start_time_seconds":"1613527390","unity_version":"2019.4.18f1","command_line_arguments":"C:\\Program Files\\Unity\\Hub\\Editor\\2019.4.18f1\\Editor\\Unity.exe -projectpath C:\/Users\/frann\/Documents\/UnityProjects\/CS467\/analysis_logging -useHub -hubIPC -cloudEnvironment production -hubSessionId c9277360-70a6-11eb-9363-db6cc44d73e6 -accessToken D6fAzro6ock-1efwinfDKR7nsoVY-KpOvxwKGxE8CQY002f","communication_protocol_version":"1.0.0","com.unity.ml-agents_version":"1.0.6","scene_name":"Level_FR","end_time_seconds":"1613527395"}}
