+++ Assets/ML-Agents/Timers/Training_0_timers.json
@@ -1,1 +1,1 @@
-{"count":1,"self":27.8885184,"total":28.5299421,"children":{"AgentSendState":{"count":1114,"self":0.057592399999999995,"total":0.0695886,"children":{"CollectObservations":{"count":223,"self":0.010989299999999999,"total":0.010989299999999999,"children":null},"CollectDiscreteActionMasks":{"count":223,"self":0.0010069,"total":0.0010069,"children":null}}},"DecideAction":{"count":1114,"self":0.5317727,"total":0.5317727,"children":null},"AgentAct":{"count":1114,"self":0.0380621,"total":0.0380621,"children":null}},"gauges":{"Paddle.CumulativeReward":{"count":4,"max":12,"min":-8,"runningAverage":-1.5,"value":-8,"weightedAverage":-3.40625}},"metadata":{"timer_format_version":"0.1.0","start_time_seconds":"1613437754","unity_version":"2019.4.18f1","command_line_arguments":"C:\\Program Files\\Unity\\Hub\\Editor\\2019.4.18f1\\Editor\\Unity.exe -projectpath C:\/Users\/frann\/Documents\/UnityProjects\/CS467\/analysis_logging -useHub -hubIPC -cloudEnvironment production -hubSessionId 35b826e0-6fef-11eb-8804-59ce1fd46b66 -accessToken BDdcbCpNTfX5zDOGnVxwAP4MYYOa0xD8GqTTGdXLwDU002f","communication_protocol_version":"1.0.0","com.unity.ml-agents_version":"1.0.6","scene_name":"Training_0","end_time_seconds":"1613437782"}}
+{"count":1,"self":7.8988064,"total":8.7517057,"children":{"AgentSendState":{"count":3762,"self":0.0736836,"total":0.0911143,"children":{"CollectObservations":{"count":753,"self":0.0100068,"total":0.0100068,"children":null},"CollectDiscreteActionMasks":{"count":753,"self":0.0074239,"total":0.0074239,"children":null}}},"DecideAction":{"count":3762,"self":0.7101004,"total":0.7101004,"children":null},"AgentAct":{"count":3762,"self":0.051684299999999996,"total":0.051684299999999996,"children":null}},"gauges":{"Paddle.CumulativeReward":{"count":15,"max":24,"min":-10,"runningAverage":0.399999917,"value":-4,"weightedAverage":0.7893492}},"metadata":{"timer_format_version":"0.1.0","start_time_seconds":"1613527881","unity_version":"2019.4.18f1","command_line_arguments":"C:\\Program Files\\Unity\\Hub\\Editor\\2019.4.18f1\\Editor\\Unity.exe -projectpath C:\/Users\/frann\/Documents\/UnityProjects\/CS467\/analysis_logging -useHub -hubIPC -cloudEnvironment production -hubSessionId c9277360-70a6-11eb-9363-db6cc44d73e6 -accessToken D6fAzro6ock-1efwinfDKR7nsoVY-KpOvxwKGxE8CQY002f","communication_protocol_version":"1.0.0","com.unity.ml-agents_version":"1.0.6","scene_name":"Training_0","end_time_seconds":"1613527890"}}
