+++ Assets/ML-Agents/Timers/Training_Multi_timers.json
@@ -1,1 +1,1 @@
-{"count":1,"self":14.246492799999999,"total":14.7468676,"children":{"AgentSendState":{"count":464,"self":0.0936354,"total":0.1405792,"children":{"CollectObservations":{"count":1488,"self":0.027804299999999997,"total":0.027804299999999997,"children":null},"CollectDiscreteActionMasks":{"count":1488,"self":0.0191395,"total":0.0191395,"children":null}}},"DecideAction":{"count":464,"self":0.319704,"total":0.319704,"children":null},"AgentAct":{"count":464,"self":0.0360939,"total":0.0360939,"children":null}},"gauges":{"Paddle.CumulativeReward":{"count":12,"max":-8.599999,"min":-19.5999985,"runningAverage":-10.1083336,"value":-9.4,"weightedAverage":-9.584768}},"metadata":{"timer_format_version":"0.1.0","start_time_seconds":"1613437795","unity_version":"2019.4.18f1","command_line_arguments":"C:\\Program Files\\Unity\\Hub\\Editor\\2019.4.18f1\\Editor\\Unity.exe -projectpath C:\/Users\/frann\/Documents\/UnityProjects\/CS467\/analysis_logging -useHub -hubIPC -cloudEnvironment production -hubSessionId 35b826e0-6fef-11eb-8804-59ce1fd46b66 -accessToken BDdcbCpNTfX5zDOGnVxwAP4MYYOa0xD8GqTTGdXLwDU002f","communication_protocol_version":"1.0.0","com.unity.ml-agents_version":"1.0.6","scene_name":"Training_Multi","end_time_seconds":"1613437810"}}
+{"count":1,"self":16.262472,"total":16.5502647,"children":{"AgentSendState":{"count":581,"self":0.0409473,"total":0.0591451,"children":{"CollectObservations":{"count":1872,"self":0.011103799999999999,"total":0.011103799999999999,"children":null},"CollectDiscreteActionMasks":{"count":1872,"self":0.007094,"total":0.007094,"children":null}}},"DecideAction":{"count":581,"self":0.2095688,"total":0.2095688,"children":null},"AgentAct":{"count":581,"self":0.0180787,"total":0.0180787,"children":null}},"gauges":{"Paddle.CumulativeReward":{"count":14,"max":-6.800001,"min":-20,"runningAverage":-9.492858,"value":-9.8,"weightedAverage":-9.659445}},"metadata":{"timer_format_version":"0.1.0","start_time_seconds":"1613527906","unity_version":"2019.4.18f1","command_line_arguments":"C:\\Program Files\\Unity\\Hub\\Editor\\2019.4.18f1\\Editor\\Unity.exe -projectpath C:\/Users\/frann\/Documents\/UnityProjects\/CS467\/analysis_logging -useHub -hubIPC -cloudEnvironment production -hubSessionId c9277360-70a6-11eb-9363-db6cc44d73e6 -accessToken D6fAzro6ock-1efwinfDKR7nsoVY-KpOvxwKGxE8CQY002f","communication_protocol_version":"1.0.0","com.unity.ml-agents_version":"1.0.6","scene_name":"Training_Multi","end_time_seconds":"1613527923"}}
