+++ performance_tracking/raw_._Paddle - bo_13_-_08.csv
@@ -1,0 +1,16 @@
+ModelName,GameNumber,NumGamesPlayed,Score,BlockHits,PaddleHits,TimePlayed,WinStatus
+Paddle - bo_13,1,15,40,4,0,101.72,0
+Paddle - bo_13,2,15,10,1,0,0.72,0
+Paddle - bo_13,3,15,20,2,1,2.09,0
+Paddle - bo_13,4,15,20,2,1,1.64,0
+Paddle - bo_13,5,15,40,4,0,3.32,0
+Paddle - bo_13,6,15,10,1,0,0.48,0
+Paddle - bo_13,7,15,10,1,0,0.94,0
+Paddle - bo_13,8,15,10,1,0,0.54,0
+Paddle - bo_13,9,15,10,1,0,0.96,0
+Paddle - bo_13,10,15,0,0,0,1.74,0
+Paddle - bo_13,11,15,60,6,0,4.79,0
+Paddle - bo_13,12,15,80,8,1,6.23,0
+Paddle - bo_13,13,15,10,1,0,2.55,0
+Paddle - bo_13,14,15,10,1,0,0.6,0
+Paddle - bo_13,15,15,70,7,2,4.65,0
