+++ performance_tracking/raw_._Paddle - bo_13_-_10.csv
@@ -1,0 +1,16 @@
+ModelName,GameNumber,NumGamesPlayed,Score,BlockHits,PaddleHits,TimePlayed,WinStatus
+Paddle - bo_13,1,15,20,2,1,101.99,0
+Paddle - bo_13,2,15,80,8,3,6.68,0
+Paddle - bo_13,3,15,70,7,0,3.97,0
+Paddle - bo_13,4,15,90,9,1,6.15,0
+Paddle - bo_13,5,15,10,1,0,0.58,0
+Paddle - bo_13,6,15,170,17,0,6.52,0
+Paddle - bo_13,7,15,20,2,1,3.09,0
+Paddle - bo_13,8,15,10,1,0,0.7,0
+Paddle - bo_13,9,15,10,1,0,0.76,0
+Paddle - bo_13,10,15,10,1,2,3.11,0
+Paddle - bo_13,11,15,0,0,0,2.43,0
+Paddle - bo_13,12,15,40,4,1,5.02,0
+Paddle - bo_13,13,15,140,14,1,21.21,0
+Paddle - bo_13,14,15,80,8,2,6.04,0
+Paddle - bo_13,15,15,30,3,1,4.43,0
