+++ performance_tracking/summary_._Paddle - bo_13_-_05.csv
@@ -1,0 +1,4 @@
+ModelName,NumGamesPlayed,Statistic,Score,BlockHits,PaddleHits,TimePlayed,WinPercentage
+Paddle - bo_13,1398,Minimum,0,0,0,0.08,0.0072
+Paddle - bo_13,1398,Average,58.68,5.87,1.45,6.97,0.0072
+Paddle - bo_13,1398,Maximum,210,21,40,100.1,0.0072
