+++ performance_tracking/summary_._Paddle - bo_13_-_08.csv
@@ -1,0 +1,4 @@
+ModelName,NumGamesPlayed,Statistic,Score,BlockHits,PaddleHits,TimePlayed,WinPercentage
+Paddle - bo_13,15,Minimum,0,0,0,0.48,0
+Paddle - bo_13,15,Average,26.67,2.67,0.33,8.86,0
+Paddle - bo_13,15,Maximum,80,8,2,101.72,0
