+++ performance_tracking/summary_._Paddle - bo_13_-_10.csv
@@ -1,0 +1,4 @@
+ModelName,NumGamesPlayed,Statistic,Score,BlockHits,PaddleHits,TimePlayed,WinPercentage
+Paddle - bo_13,15,Minimum,0,0,0,0.58,0
+Paddle - bo_13,15,Average,52,5.2,0.87,11.51,0
+Paddle - bo_13,15,Maximum,170,17,3,101.99,0
