+++ performance_tracking/summary_.__-_09.csv
@@ -1,0 +1,4 @@
+ModelName,NumGamesPlayed,Statistic,Score,BlockHits,PaddleHits,TimePlayed,WinPercentage
+,0,Minimum,0,0,0,0,0
+,0,Average,NaN,NaN,NaN,NaN,0
+,0,Maximum,0,0,0,0,0
